converts form components to typescript
