+++ src/components/form/field_text/index.ts
@@ -1,1 +1,1 @@
-export { EuiFieldText } from './field_text';
+export { EuiFieldText, EuiFieldTextProps } from './field_text';
