+++ src/components/form/form.tsx
@@ -1,20 +1,33 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import React, { FC, FormHTMLAttributes, ReactNode } from 'react';
 import classNames from 'classnames';
 import { EuiCallOut } from '../call_out';
 import { EuiI18n } from '../i18n';
+import { CommonProps } from '../common';
+import makeId from './form_row/make_id';
 
-export const EuiForm = ({ children, className, isInvalid, error, ...rest }) => {
+export type EuiFormProps = CommonProps &
+  FormHTMLAttributes<HTMLFormElement> & {
+    isInvalid?: boolean;
+    error?: ReactNode | ReactNode[];
+  };
+
+export const EuiForm: FC<EuiFormProps> = ({
+  children,
+  className,
+  isInvalid,
+  error,
+  ...rest
+}) => {
   const classes = classNames('euiForm', className);
 
-  let optionalErrors;
+  let optionalErrors: JSX.Element | null = null;
 
   if (error) {
     const errorTexts = Array.isArray(error) ? error : [error];
     optionalErrors = (
       <ul>
         {errorTexts.map(error => (
-          <li className="euiForm__error" key={error}>
+          <li className="euiForm__error" key={makeId()}>
             {error}
           </li>
         ))}
@@ -29,7 +42,7 @@ export const EuiForm = ({ children, className, isInvalid, error, ...rest }) => {
       <EuiI18n
         token="euiForm.addressFormErrors"
         default="Please address the errors in your form.">
-        {addressFormErrors => (
+        {(addressFormErrors: string) => (
           <EuiCallOut
             className="euiForm__errors"
             title={addressFormErrors}
@@ -42,17 +55,9 @@ export const EuiForm = ({ children, className, isInvalid, error, ...rest }) => {
   }
 
   return (
-    <div className={classes} {...rest}>
+    <form className={classes} {...rest}>
       {optionalErrorAlert}
       {children}
-    </div>
+    </form>
   );
-};
-
-EuiForm.propTypes = {
-  isInvalid: PropTypes.bool,
-  error: PropTypes.oneOfType([
-    PropTypes.node,
-    PropTypes.arrayOf(PropTypes.node),
-  ]),
 };
