+++ src/components/form/form_control_layout/index.ts
@@ -10,4 +10,7 @@ export {
   EuiFormControlLayoutCustomIcon,
 } from './form_control_layout_custom_icon';
 
-export { EuiFormControlLayoutDelimited } from './form_control_layout_delimited';
+export {
+  EuiFormControlLayoutDelimited,
+  EuiFormControlLayoutDelimitedProps,
+} from './form_control_layout_delimited';
