+++ src/components/form/form_error_text/form_error_text.tsx
@@ -1,11 +1,16 @@
-import React, { FunctionComponent, HTMLAttributes } from 'react';
+import React, { FC, HTMLAttributes } from 'react';
 
 import classNames from 'classnames';
 import { CommonProps } from '../../common';
 
-export const EuiFormErrorText: FunctionComponent<
-  CommonProps & HTMLAttributes<HTMLDivElement>
-> = ({ children, className, ...rest }) => {
+export type EuiFormErrorTextProps = CommonProps &
+  HTMLAttributes<HTMLDivElement>;
+
+export const EuiFormErrorText: FC<EuiFormErrorTextProps> = ({
+  children,
+  className,
+  ...rest
+}) => {
   const classes = classNames('euiFormErrorText', className);
 
   return (
