+++ src/components/form/form_help_text/index.ts
@@ -1,1 +1,1 @@
-export { EuiFormHelpText } from './form_help_text';
+export { EuiFormHelpText, EuiFormHelpTextProps } from './form_help_text';
