--- src/components/form/index.js
@@ -1,24 +1,0 @@
-export { EuiCheckbox, EuiCheckboxGroup } from './checkbox';
-export { EuiDescribedFormGroup } from './described_form_group';
-export { EuiFieldNumber } from './field_number';
-export { EuiFieldPassword } from './field_password';
-export { EuiFieldSearch } from './field_search';
-export { EuiFieldText } from './field_text';
-export { EuiFilePicker } from './file_picker';
-export { EuiForm } from './form';
-export {
-  EuiFormControlLayout,
-  EuiFormControlLayoutDelimited,
-} from './form_control_layout';
-export { EuiFormErrorText } from './form_error_text';
-export { EuiFormFieldset, EuiFormLegend } from './form_fieldset';
-export { EuiFormHelpText } from './form_help_text';
-export { EuiFormLabel } from './form_label';
-export { EuiFormRow } from './form_row';
-export { EuiRadio, EuiRadioGroup } from './radio';
-export { EuiDualRange, EuiRange } from './range';
-export { EuiSelect } from './select';
-export { EuiSuperSelect, EuiSuperSelectControl } from './super_select';
-export { EuiSwitch } from './switch';
-export { EuiTextArea } from './text_area';
-export { EuiValidatableControl } from './validatable_control';
