+++ src/components/form/validatable_control/index.ts
@@ -1,1 +1,4 @@
-export { EuiValidatableControl } from './validatable_control';
+export {
+  EuiValidatableControl,
+  EuiValidatableControlProps,
+} from './validatable_control';
