+++ src/components/index.d.ts
@@ -1,5 +1,4 @@
 /// <reference path="./date_picker/index.d.ts" />
-/// <reference path="./form/index.d.ts" />
 
 declare module '@elastic/eui' {
   // @ts-ignore
